requirements.txt
django
django-appconf>=0.6
